Enterprise Inventory Warehouse management  - Warehouse CRUD  - Option to set default warehouse  - Event listener to force only one default warehouse

+++ src/MarelloEnterprise/Bundle/InventoryBundle/Controller/WarehouseController.php
@@ -3,10 +3,14 @@
 namespace MarelloEnterprise\Bundle\InventoryBundle\Controller;
 
 use Marello\Bundle\InventoryBundle\Entity\Warehouse;
+use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
+use Oro\Bundle\SecurityBundle\Authentication\Token\UsernamePasswordOrganizationToken;
+use Oro\Bundle\UserBundle\Entity\User;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration as Config;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Zend\Stdlib\Response;
+use Symfony\Component\HttpFoundation\Response;
 
 /**
  * @Config\Route("/warehouse")
@@ -17,6 +21,7 @@ class WarehouseController extends Controller
      * @Config\Route("/")
      * @Config\Method("GET")
      * @Config\Template
+     * @AclAncestor("marello_inventory_warehouse_view")
      */
     public function indexAction()
     {
@@ -28,29 +33,38 @@ class WarehouseController extends Controller
     /**
      * @Config\Route("/create")
      * @Config\Method({"GET", "POST"})
+     * @Config\Template("@MarelloEnterpriseInventory/Warehouse/update.html.twig")
+     * @AclAncestor("marello_inventory_warehouse_create")
      *
-     * @param Request   $request
+     * @param Request $request
+     *
+     * @return RedirectResponse
      */
     public function createAction(Request $request)
     {
-        // TODO: Create
+        return $this->update($request);
     }
 
     /**
      * @Config\Route("/update/{id}")
      * @Config\Method({"GET", "POST"})
+     * @Config\Template
+     * @AclAncestor("marello_inventory_warehouse_update")
      *
      * @param Warehouse $warehouse
      * @param Request   $request
+     *
+     * @return RedirectResponse
      */
     public function updateAction(Warehouse $warehouse, Request $request)
     {
-        // TODO: Update
+        return $this->update($request, $warehouse);
     }
 
     /**
      * @Config\Route("/delete/{id}")
      * @Config\Method("DELETE")
+     * @AclAncestor("marello_inventory_warehouse_delete")
      *
      * @param Warehouse $warehouse
      *
@@ -60,9 +74,60 @@ class WarehouseController extends Controller
     {
         $em = $this->getDoctrine()->getManager();
 
+        if ($warehouse->isDefault()) {
+            // TODO: Cannot remove default warehouse.
+            $this->addFlash('error', 'Cannot delete default warehouse.');
+
+            return new Response('Cannot delete default warehouse.', Response::HTTP_INTERNAL_SERVER_ERROR);
+        }
+
         $em->remove($warehouse);
         $em->flush();
 
         return new Response();
+    }
+
+    /**
+     * @param Request        $request
+     * @param Warehouse|null $warehouse
+     *
+     * @return RedirectResponse
+     */
+    protected function update(Request $request, Warehouse $warehouse = null)
+    {
+        if (!$warehouse) {
+            $warehouse = new Warehouse();
+
+            /** @var UsernamePasswordOrganizationToken $token */
+            $token = $this->get('security.token_storage')->getToken();
+
+            $warehouse->setOwner($token->getOrganizationContext());
+        }
+
+        $form = $this->createForm('marello_warehouse', $warehouse);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+
+            $em->persist($warehouse);
+            $em->flush();
+
+            return $this->get('oro_ui.router')->redirectAfterSave(
+                [
+                    'route'      => 'marelloenterprise_inventory_warehouse_update',
+                    'parameters' => ['id' => $warehouse->getId()],
+                ],
+                [
+                    'route' => 'marelloenterprise_inventory_warehouse_index',
+                ],
+                $warehouse
+            );
+        }
+
+        return [
+            'entity' => $warehouse,
+            'form'   => $form->createView(),
+        ];
     }
 }
